pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"


[project]
name = "tagfile"
authors = [{name = "Benjamin Althues", email = "benjamin@babab.nl"}]
readme = "README.rst"
requires-python = ">=3.7"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta",
    #"Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Database",
    "Topic :: Desktop Environment :: File Managers",
    "Topic :: Multimedia",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: System :: Filesystems",
    "Topic :: Utilities",
]
keywords = [
    "tag", "files", "media", "duplicate", "duplicates", "organize",
    "checksum", "locate", "database", "find", "hash", "filter",
    "harddisk", "space"
]
dynamic = ["version", "description"]
dependencies = [
    "ansicolors >=1.1.8,<2.0",
    "peewee >=3.15.4,<4.0",
    "pycommand >=0.4.0,<0.5",
    "PyYAML >=6.0,<7.0",
]


[project.urls]
Home = "https://github.com/babab/tagfile"


[project.scripts]
tagfile = "tagfile:main"


[tool.flit.sdist]
exclude = [".circleci", ".gitignore", "Makefile"]
